TextureSamplerUtil の整理
